update to net 6.. all seems ot work .. added a x64 config.. left core as any cpu in case ARM windows gets supported..
not sure if all the config could by anycpu but got warningins about arm...it all work the problem is, i dont use the GameComponent, i prefer teh other way and juse use the Device.. its not needed to have a game class, and in my use case i handle input via WPF...  i was usng the MGcontent control but it lacked all the workarounds and thiings in here for tabbed devices, i put in my own fixes but now i cant get it to work on net 6 wpf.   so i might try this or try to factor out the game refs to juse device service. or its not a trivial merge.. also there is a https://github.com/microsoft/WPFDXInterop which hasnt been update in ages.. i dont know if im going to move to avalonia or maui or something based on monogame butg i have years invested in legacy wpf editor, and its wroking on wiondows... I would love if thsi stuff woks on ARM surface go I can develop androids apps on one, using vs 20222 arm an DX... right now im wating for this to pan out.. not sure if this is a worth PR and i bumped the version to 2 because its not netstand 2.0 compatible, its doenst work on .48  , and lots of legacy code will break if updating to MG 3.81.  so i got through it and mhy framerate is faster so it was worth updateing all my cores to net6 libs...even if wpf netcore6 only works on windows 64 and maybe 32.   theres commnets taht 32 bit builds are acutally faster... in manyh situations so i left it in.  anyways if you dont care to merge , i wont be offended.. i just need to try to get this back in my tool somehow, mabye ill merge some of your approach with the Wpfcontent control by craftwork games.  ( i updated his sample but that works on ly in a simple app no tabs)   i get an GetSharedHandle error creating render target im stuck there.   anways hope this helps mabye it could be a branch..